Refactor settings tests and add edge case test
- Split SettingsUseCasesTest into separate files per use case
- Use shared FakePreferencesRepository from test/Fakes.kt
- Add edge case test for daily memo without timestamps
- Improve code reuse and test organization

Coverage: 86.91% → 87.20% (+6 lines)
