+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/settings/domain/usecase/SaveTimeRangeTabUseCaseTest.kt
@@ -1,0 +1,52 @@
+package space.be1ski.vibits.shared.feature.settings.domain.usecase
+
+import space.be1ski.vibits.shared.feature.settings.domain.model.AppLanguage
+import space.be1ski.vibits.shared.feature.settings.domain.model.AppTheme
+import space.be1ski.vibits.shared.feature.settings.domain.model.TimeRangeTab
+import space.be1ski.vibits.shared.feature.settings.domain.model.UserPreferences
+import space.be1ski.vibits.shared.test.FakePreferencesRepository
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class SaveTimeRangeTabUseCaseTest {
+  @Test
+  fun `when save HABITS screen tab then updates habitsTimeRangeTab`() {
+    val repository = FakePreferencesRepository()
+    val useCase = SaveTimeRangeTabUseCase(repository)
+
+    useCase(TimeRangeScreen.HABITS, TimeRangeTab.MONTHS)
+
+    assertEquals(TimeRangeTab.MONTHS, repository.stored.habitsTimeRangeTab)
+    assertEquals(TimeRangeTab.WEEKS, repository.stored.postsTimeRangeTab)
+  }
+
+  @Test
+  fun `when save POSTS screen tab then updates postsTimeRangeTab`() {
+    val repository = FakePreferencesRepository()
+    val useCase = SaveTimeRangeTabUseCase(repository)
+
+    useCase(TimeRangeScreen.POSTS, TimeRangeTab.QUARTERS)
+
+    assertEquals(TimeRangeTab.WEEKS, repository.stored.habitsTimeRangeTab)
+    assertEquals(TimeRangeTab.QUARTERS, repository.stored.postsTimeRangeTab)
+  }
+
+  @Test
+  fun `when save then preserves other preferences`() {
+    val initialPrefs =
+      UserPreferences(
+        habitsTimeRangeTab = TimeRangeTab.WEEKS,
+        postsTimeRangeTab = TimeRangeTab.MONTHS,
+        theme = AppTheme.DARK,
+        language = AppLanguage.RUSSIAN,
+      )
+    val repository = FakePreferencesRepository(initialPrefs)
+    val useCase = SaveTimeRangeTabUseCase(repository)
+
+    useCase(TimeRangeScreen.HABITS, TimeRangeTab.YEARS)
+
+    assertEquals(TimeRangeTab.YEARS, repository.stored.habitsTimeRangeTab)
+    assertEquals(AppTheme.DARK, repository.stored.theme)
+    assertEquals(AppLanguage.RUSSIAN, repository.stored.language)
+  }
+}
